+++ frontend/src/components/logins/login-signup.css
@@ -2,10 +2,10 @@
     display: flex;
     flex-direction: column;
     margin: auto;
-    margin-top: 200px;
-    width: 600px;
+    margin-top: 190px;
+    width: 520px;
     background-color: #fff;
-    padding-bottom: 30px;
+    padding-bottom: 20px;
 }
 
 .header {
@@ -19,7 +19,7 @@
 
 .text {
     color: #086591;
-    font-size: 30px;
+    font-size: 45px;
     font-weight: 600;
 }
 
@@ -30,6 +30,7 @@
     border-radius: 10px;
 }
 
+
 .inputs {
     margin-top: 55px;
     display: flex;
@@ -37,29 +38,31 @@
     gap: 25px;
 }
 
+
 .input {
     display: flex;
     align-items: center;
     margin: auto;
-    width: 480px;
-    height: 80px;
+    width: 450px;
+    height: 60px;
     background: #f5f5f5;
     border-radius: 6px;
 }
 
 .input input {
-    width: 400px;
-    height: 40px;
+    height: 30px;
+    width: 350px;
     outline: none;
     border: none;
     background: transparent;
-    font-size: 19px;
-    font-weight: 400;
+    font-size: 15px;
     color: #797979;
+    margin: 0 15px;
 }
 
+
 .forgot-password {
-    padding-left: 62px;
+    padding-left: 40px;
     margin-top: 27px;
     color: #797979;
     font-size: 18px;
@@ -72,20 +75,25 @@
 
 .submit-container {
     display: flex;
-    gap: 30px;
-    margin-top: 50px auto;
+    gap: 20px;
+    margin: 40px auto;
 }
 
 .submit {
     display: flex;
     justify-content: center;
     align-items: center;
-    width: 200px;
-    height: 50px;
+    width: 190px;
+    height: 40px;
     color: #fff;
-    font-size: 20px;
+    font-size: 15px;
     background: #086591;
-    border-radius: 50px;
+    border-radius: 30px;
     cursor: pointer;
-    font-weight: 600;
+    font-weight: 300;
+}
+
+.gray {
+    background: #f5f5f5;
+    color: #797979;
 }
